Fix 32-bit precalc

+++ src/player/player.cc
@@ -238,9 +238,9 @@ PlaybackResult playback_loop(
     bool seeked = false;
     const size_t bytespersec = 4 * state.frequency;
     // UADE plays some mods for hours or possibly forever (with always_ends default)
-    size_t maxbytes = config.known_timeout > 0 ?
+    uint64_t maxbytes = config.known_timeout > 0 ?
         config.known_timeout * bytespersec / 1000 : PRECALC_TIMEOUT * bytespersec;
-    size_t totalbytes = 0;
+    uint64_t totalbytes = 0;
 
     while (!(stopped = check_stop()) && totalbytes < maxbytes) {
         int seek_millis = check_seek();
